move observe.go to job package
