+++ internal/search/job/observe.go
@@ -1,4 +1,4 @@
-package jobutil
+package job
 
 import (
 	"context"
